IDEA-162157 Stream API migration: support counted for loop (both outer and flatMapped)

+++ java/java-tests/testData/codeInsight/daemonCodeAnalyzer/quickFix/streamApiMigration/beforePrimitiveArrayWrongType.java
@@ -1,7 +1,7 @@
-// "Replace with forEach" "true"
+// "Replace with forEach" "false"
 public class Main {
   public void test(int[] arr) {
-    for(int i : a<caret>rr) {
+    for(float i : a<caret>rr) {
       if(i > 0) {
         System.out.println(i);
       }
